Add files via upload
This program emulates the use of multiple different Python methods/functions on a list (a two dimensional array of strings), described by the starting menu displayed. The file input.txt is an example list that may be used with the program, any other text file with less than 256 characters per line would work as well.